backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.26.2
werkzeug==3.0.1
ollama==0.4.4
